+++ sdk/keyvault/azure-security-keyvault-keys/src/key_client.cpp
@@ -45,13 +45,13 @@ static inline RequestWithContinuationToken BuildRequestFromContinuationToken(
     request.Path.clear();
     request.Path.emplace_back(nextPageUrl.GetPath());
   }
-  if (options.MaxResults)
+  if (options.MaxPageResults)
   {
     if (request.Query == nullptr)
     {
       request.Query = std::make_unique<std::map<std::string, std::string>>();
     }
-    request.Query->emplace("maxResults", std::to_string(options.MaxResults.Value()));
+    request.Query->emplace("maxResults", std::to_string(options.MaxPageResults.Value()));
   }
   return request;
 }