+++ sdk/keyvault/azure-security-keyvault-keys/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Removed `Azure::Security::KeyVault::Keys::JsonWebKey::to_json`.
 - Replaced static functions from `KeyOperation` and `KeyCurveName` for static const members.
 - Replaced the enum `JsonWebKeyType` for a class with static const members as ext enum.
+- Renamed `MaxResults` to `MaxPageResults` for `GetSinglePageOptions`.
 
 ## 4.0.0-beta.1 (2021-04-07)